Include repository name in badge URI
I’ve added a `repo` URI parameter to the example code for the Volkswagen badge URI, to ensure that the correct status for a given project is reported. After all, nobody wants to accidentally report the CI status of a different project.

In the future, this could be extended to projects not hosted on GitHub.